rename index_sorter to doc_id_mapping
use bufferlender in recorder
